Fix the build and release tasks to do both MRI and JRuby
The build task will build both the MRI and JRuby gem, and the release task will release both gems. The combined install would fail, though, so it's not included.

The build task depends on lib/snappy_ext.jar to make sure that that file is included in the JRuby gem. It needs to be built from JRuby, but the build task doesn't have to be run from JRuby.